refactor: move handle_result to cli module
